requirements.txt
flask
pyserial
mysql-connector-python
pytz